android native lib

+++ FaustGames/Native/src/core/Environment.h
@@ -8,7 +8,7 @@ namespace core
 	class Environment
 	{
 	public:
-		const static uint LongTimeScale = 0x100000000; // does 136 years of game time good enough?
+		const static ulong LongTimeScale = 0x100000000; // does 136 years of game time good enough?
 		const static uint LongFramesScale = 0x8000000;
 		static ulong gameTime;
 		static ulong elapsedTime;
